requirements.txt
django
psycopg2
pillow
djangorestframework
python-dotenv
django-filter
djangorestframework-simplejwt
coverage